+++ CHANGELOG.md
@@ -38,6 +38,7 @@ All notable changes to this project will be documented in this file.
 * CloneClassAsSharedRef missing the Attribute suffix. This is technically breaking but it'd be really weird to rely on the exact class name.
 * Obsolete documentation on StaticReferencesAttribute.Initialized(), along with a related anti-optimization.
 * User-type discovery could crash when a user assembly had a broken dependency.
+* Changing `Config.ConverterFactory` would not update caches properly.
 
 ## [v0.11.0]
 ### Added